backend/src/first_req.py


import dotenv
from google import genai
dotenv.load_dotenv()


client = genai.Client()

response = client.models.generate_content(
    model="gemini-2.5-flash-image",
    contents="Create a picture of a nano banana dish in a fancy restaurant with a Gemini theme"
)


for part in response.parts:
    if part.text is not None:
        print(part.text)
    elif part.inline_data is not None:
        image = part.as_image()
        image.save("generated_image.png")